Conversor monetario aparcao

+++ frontend/BeCitizeN/app/src/main/java/com/becitizen/app/becitizen/domain/ControllerUtilitiesDomain.java
@@ -18,7 +18,7 @@ public class ControllerUtilitiesDomain {
         ControllerUtititesData.getInstance().getCurrencyList(currencyList);
     }
 
-    public int getConversion(String currencyFrom, String currencyTo, String amount) {
+    public double getConversion(String currencyFrom, String currencyTo, String amount) {
         return ControllerUtititesData.getInstance().getConversion(currencyFrom,currencyTo,amount);
     }
 }
